Moved EventData interface to types module
